update
item layout in check-ins was fixed

Activity feed added

in the activity list selecting an activity will take you to the venues
page

the venues page has a couple of interesting features. instead of having
separate pages for photos and tips, everything is on one page. Photos
and tips are horizontally scroll. so to see more photos and/or tips
scroll left and right. To see the rest of the information about the
venue like the mayor and rating scroll down.